refactor(notes): rename presenter to SurveyDatePicker, modal to IosSurveyDatePicker
The lib presenter is what the app talks to, so it takes the clean name
SurveyDatePicker (require 'logic/SurveyDatePicker'); the iOS-only inline-
calendar modal controller becomes IosSurveyDatePicker, removing the
name collision. Also register IosSurveyDatePicker in the spec runner's
SPEC_FILES — the modal spec was never listed, so it had never run.
