+++ walta-app/app/lib/android/logic/SurveyDatePicker.js
