+++ walta-app/app/lib/ios/logic/SurveyDatePicker.js
@@ -1,8 +1,8 @@
 // iOS has no native modal date dialog, so the survey date is chosen through the
-// SurveyDatePicker overlay (an inline calendar). open() returns a handle the
+// IosSurveyDatePicker overlay (an inline calendar). open() returns a handle the
 // caller holds to tear the modal down; datePicker is exposed for introspection.
 function open({ window, date, onSelected, onClose }) {
-    var modal = Alloy.createController("SurveyDatePicker", { date });
+    var modal = Alloy.createController("IosSurveyDatePicker", { date });
     window.add( modal.getView() );
     modal.on("selected", ( e ) => onSelected( e.value ));
     modal.on("close", onClose);
