+++ walta-app/app/styles/IosSurveyDatePicker.tss
