+++ walta-app/app/views/IosSurveyDatePicker.xml
